Added global change view
